backend/requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.4.0

# Database
psycopg[binary]==3.3.2

# Environment variables
python-decouple==3.8

# Development
django-extensions==3.2.3

# Production
gunicorn==23.0.0
whitenoise==6.8.2
